Prefer master nodes for running the controller
For the all-in-one case, the edge controller manager has to be run
in the master nodes. For services with low traffic volume, this is
also acceptable, so it can be considered the default case since for
high traffic cases a true load balancer should be used anyways.

This commit add a toleration to the NoSchedule on master nodes and
also add a node selector to select them.