Remove nodeSelector
Remove the master node selector so that it can run
on deployments without master-labeled nodes.
(like k3s with --disable-agent.)

From the commit log, I don't see why this selector was necessary
in the first place.  The all-in-one case mentioned in the commit
message should be fine without this selector.